devenv/mgit/requirements.txt
click==6.7
PyYaml==3.12
